shop

+++ core/src/main/java/com/avaricious/components/automations/SpinBuyerSpeed.java
@@ -17,7 +17,7 @@ public class SpinBuyerSpeed extends AbstractAutomationUpgrade {
     }
 
     @Override
-    public int price() {
-        return 0;
+    public float price() {
+        return 0f;
     }
 }
