+++ core/src/main/java/com/avaricious/components/popups/CreditNumberPopup.java
@@ -12,11 +12,11 @@ public class CreditNumberPopup extends NumberPopup {
 
     private final TextureRegion dollarSymbol = Assets.I().get(AssetKey.DOLLAR_SYMBOL);
 
-    public CreditNumberPopup(int number, float x, float y, boolean asPercentage, boolean manualHold) {
+    public CreditNumberPopup(float number, float x, float y, boolean asPercentage, boolean manualHold) {
         super(number, Assets.I().yellow(), x, y, asPercentage, manualHold);
     }
 
-    public CreditNumberPopup(int number, Rectangle bounds, boolean asPercentage, boolean manualHold) {
+    public CreditNumberPopup(float number, Rectangle bounds, boolean asPercentage, boolean manualHold) {
         super(number, Assets.I().yellow(), bounds, asPercentage, manualHold);
     }
